fixed e12872b from https://gitee.com/sunjiakun/arkui_ace_engine/pulls/16781 fix calendar picker js cast issue

+++ frameworks/core/components_ng/pattern/calendar/calendar_pattern.cpp
@@ -350,8 +350,9 @@ void CalendarPattern::FlushDialogMonthData(ObtainedMonth& obtainedMonth)
         return;
     }
     for (auto& day : obtainedMonth.days) {
-        day.isSelected = day.month.year == selectedDay_.GetYear() && day.month.month == selectedDay_.GetMonth() &&
-                       day.day == selectedDay_.GetDay();
+        day.isSelected = day.month.year == static_cast<int32_t>(selectedDay_.GetYear()) &&
+                         day.month.month == static_cast<int32_t>(selectedDay_.GetMonth()) &&
+                         day.day == static_cast<int32_t>(selectedDay_.GetDay());
     }
 }
